refactor(core): split frame stats and assemble render model
Break frame.ts into helpers, non-identity stats, and candidate factories.
Extract domain snapshot computation and RenderModel assembly so runPipeline
stays a short phase chain. Tests cover dispose/row contracts and frozen
domain snapshots.
