+++ packages/core/src/pipeline/frame-helpers.ts
@@ -1,0 +1,89 @@
+/**
+ * Shared frame-building helpers: empty extras, intercept lists, grouping,
+ * carried columns, and stat drop warnings.
+ */
+import { deriveGroups } from "../grouping.js";
+import type { CellValue, Discreteness } from "../table.js";
+import type { ColumnTable } from "../table.js";
+
+import type { LayerBinding, LayerFrame, PipelineWarning } from "./types.js";
+
+/** Fresh all-null frame extras (each stat branch fills what it uses). */
+export function emptyFrameExtras(): Pick<
+  LayerFrame,
+  | "ymin"
+  | "ymax"
+  | "xmin"
+  | "xmax"
+  | "dodgeSlot"
+  | "dodgeSlotCounts"
+  | "offsetX"
+  | "offsetY"
+  | "box"
+  | "smoothBand"
+  | "xIntercepts"
+  | "yIntercepts"
+> {
+  return {
+    ymin: null,
+    ymax: null,
+    xmin: null,
+    xmax: null,
+    dodgeSlot: null,
+    dodgeSlotCounts: null,
+    offsetX: null,
+    offsetY: null,
+    box: null,
+    smoothBand: false,
+    xIntercepts: [],
+    yIntercepts: [],
+  };
+}
+
+export function interceptList(value: unknown): CellValue[] {
+  if (value === undefined) return [];
+  if (Array.isArray(value)) return value as CellValue[];
+  return [value as CellValue];
+}
+
+export function deriveLayerGroups(binding: LayerBinding, table: ColumnTable): number[] {
+  const aes = binding.layer.aes ?? {};
+  const declared: Record<string, Discreteness> = {};
+  for (const mapping of Object.values(aes)) {
+    if (
+      mapping !== null &&
+      mapping !== undefined &&
+      "field" in mapping &&
+      table.has(mapping.field)
+    ) {
+      declared[mapping.field] = table.discreteness(mapping.field);
+    }
+  }
+  return [...deriveGroups(table.columns(), aes, declared).groups];
+}
+
+/** Carried discrete columns for stats (color/fill/label, minus the x field). */
+export function carriedColumns(
+  binding: LayerBinding,
+  table: ColumnTable,
+): Record<string, readonly CellValue[]> {
+  const carried: Record<string, readonly CellValue[]> = {};
+  for (const field of [binding.color.field, binding.fill.field, binding.labelField]) {
+    if (field !== null && field !== binding.xField) carried[field] = table.column(field);
+  }
+  return carried;
+}
+
+export function removedStatWarning(
+  dropped: number,
+  index: number,
+  what: string,
+  warnings: PipelineWarning[],
+): void {
+  if (dropped > 0) {
+    warnings.push({
+      code: "removed-missing",
+      message: `Removed ${dropped} row(s) with ${what} (layer ${index}).`,
+    });
+  }
+}
